refactor: update files where local storage service was used
