+++ src/features/authentication/hooks/use-log-in.tsx
@@ -1,8 +1,8 @@
 import { useMutation, UseMutationResult } from "react-query";
 import { ILogInDto, ILogInResponse } from "../dto";
 import { logIn } from "../services";
-import { setItem, StorageKeys } from "@/services/local-storage";
 import { AxiosError } from "axios";
+import { setItem, StorageKeys } from "@/services/cookies";
 
 export const useLogIn = (): UseMutationResult<
   ILogInResponse,
@@ -10,11 +10,13 @@ export const useLogIn = (): UseMutationResult<
   ILogInDto,
   unknown
 > => {
+  const THIRTY_DAYS = 30;
+
   return useMutation<ILogInResponse, AxiosError, ILogInDto, unknown>({
     mutationFn: ({ username, password }: ILogInDto) =>
       logIn({ username, password }),
     onSuccess: ({ accessToken }: ILogInResponse) => {
-      setItem(StorageKeys.COGNITO_ACCESS_TOKEN, accessToken);
+      setItem(StorageKeys.COGNITO_ACCESS_TOKEN, accessToken, THIRTY_DAYS);
     },
   });
 };
