+++ src/features/authentication/services/get-me.service.ts
@@ -1,10 +1,10 @@
 import { IGetMeResponse } from "../dto";
 import { env } from "@/config/env";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 import axios from "axios";
 
 export const getMe = async (): Promise<IGetMeResponse> => {
-  const authToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const authToken = getCognitoToken();
   const apiUrl = `${env.coreApi.baseUrl}/user/me`;
 
   const response = await axios.get<IGetMeResponse>(apiUrl, {
@@ -16,4 +16,3 @@ export const getMe = async (): Promise<IGetMeResponse> => {
 
   return response.data;
 };
-
