+++ src/features/edit-profile/services/update-profile.service.ts
@@ -2,12 +2,12 @@ import axios from "axios";
 import { env } from "@/config/env";
 import { IUpdateProfileDto } from "../interfaces";
 import { IGetMeResponse } from "@/features/authentication/dto";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 
 export const updateProfile = async (
   updateProfileDto: IUpdateProfileDto
 ): Promise<IGetMeResponse> => {
-  const authToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const authToken = getCognitoToken();
 
   const apiUrl = `${env.coreApi.baseUrl}/user/me`;
 
@@ -20,4 +20,3 @@ export const updateProfile = async (
 
   return response.data;
 };
-
