+++ src/features/stats/hooks/use-get-track-stats.ts
@@ -4,7 +4,7 @@ import {
   IMapTrackStatsToMoodTrackingResponse,
 } from "../interfaces";
 import { AxiosError } from "axios";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 import { getTrackStats } from "../services";
 import {
   mapCurrentMonthMoodTrackingData,
@@ -20,7 +20,7 @@ export const useGetTrackStats = (): UseQueryResult<
   IMapTrackStatsToMoodTrackingResponse | undefined,
   AxiosError
 > => {
-  const accessToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const accessToken = getCognitoToken();
 
   return useQuery<
     IGetTrackStatsResponse,
