+++ src/features/stats/services/get-track-stats.service.ts
@@ -1,12 +1,12 @@
 import { env } from "@/config/env";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 import axios from "axios";
 import { IGetTrackStatsResponse } from "../interfaces";
 import { getStatsStartAndEndDate } from "../utils";
 
 export const getTrackStats = async (): Promise<IGetTrackStatsResponse> => {
   const { startDate, endDate } = getStatsStartAndEndDate();
-  const authToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const authToken = getCognitoToken();
   const apiUrl = `${env.coreApi.baseUrl}/track/stats?startDate=${startDate}&endDate=${endDate}`;
 
   const response = await axios.get<IGetTrackStatsResponse>(apiUrl, {
