+++ src/features/track/services/create-track.service.ts
@@ -1,12 +1,12 @@
 import axios from "axios";
 import { env } from "@/config/env";
 import { ICreateTrackDto, ITrack } from "../interfaces";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 
 export const createTrack = async (
   createTrackDto: ICreateTrackDto
 ): Promise<ITrack> => {
-  const authToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const authToken = getCognitoToken();
 
   const apiUrl = `${env.coreApi.baseUrl}/track`;
