+++ src/features/track/services/delete-track.service.ts
@@ -1,9 +1,9 @@
 import axios from "axios";
 import { env } from "@/config/env";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 
 export const deleteTrack = async (id: number): Promise<void> => {
-  const authToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const authToken = getCognitoToken();
 
   const apiUrl = `${env.coreApi.baseUrl}/track/${id}`;
