+++ src/features/track/services/get-track-by-date-range.service.ts
@@ -1,5 +1,5 @@
 import { env } from "@/config/env";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 import axios from "axios";
 import { IGetTrackByDateRangeDto, ITrack } from "../interfaces";
 
@@ -7,7 +7,7 @@ export const getTrackByDateRange = async ({
   startDate,
   endDate,
 }: IGetTrackByDateRangeDto): Promise<ITrack[]> => {
-  const authToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const authToken = getCognitoToken();
   const apiUrl = `${env.coreApi.baseUrl}/track/by-date-range?startDate=${startDate}&endDate=${endDate}`;
 
   const response = await axios.get<ITrack[]>(apiUrl, {
