+++ src/features/track/services/update-track.service.ts
@@ -1,13 +1,13 @@
 import axios from "axios";
 import { env } from "@/config/env";
 import { ITrack, IUpdateTrackDto } from "../interfaces";
-import { getItem, StorageKeys } from "@/services/local-storage";
+import { getCognitoToken } from "@/services/cookies";
 
 export const updateTrack = async (
   id: number,
   updateTrackDto: IUpdateTrackDto
 ): Promise<ITrack> => {
-  const authToken = getItem(StorageKeys.COGNITO_ACCESS_TOKEN);
+  const authToken = getCognitoToken();
 
   const apiUrl = `${env.coreApi.baseUrl}/track/${id}`;
